strictObjectIDCoercion on role-mapping
How the json file is, loopback developers are currently unable to query RoleMappings by principalId, since most connectors (mongodb in my case) will overwrite the type of the principalId for the query to an ObjectId, which is incorrect behavior.

principalId is a string, as per this definition. So having the query cast to an ObjectId always returns no matches.